Moving the data stream reindex task to x-pack (#117927)
